+++ CHANGELOG.md
@@ -1,10 +1,12 @@
 ## [Unreleased]
 ### Added
+- Unit test via [Balatest](https://github.com/BakersDozenBagels/Balatest/) (#13)
 
 ### Changed
-
+- READ.ME to add instructions for [Balatest](https://github.com/BakersDozenBagels/Balatest/) (#13)
 ### Fixed
 - Fixed Blueprint/Brainstorm being parsed incorrectly when the mod Blueprint is installed. (#5)
+- GameState not removing a joker when Madness is activated (#15)
 
 ### Removed